requirements.txt
# Core dependencies
numpy>=1.24.0
sympy>=1.12

# Gemini API integration
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Development/Testing
pytest>=6.0.0